requirements.txt
# Основные зависимости для FastAPI
fastapi==0.100.0
uvicorn==0.22.0

# Для работы с Pydantic и валидацией данных
pydantic==1.10.7

# Для работы с капчей через 2Captcha
requests==2.28.2

# Для работы с тестированием
pytest==7.2.2
pytest-cov==3.0.0

# Для работы с Selenium
selenium==4.8.2

# Для работы с переменными окружения
python-dotenv==0.21.0
